refactor: extract phone number utilities to separate module
- Create phone-utils.ts with formatPhoneNumber and validatePhoneNumber
- Remove duplicate phone formatting logic from PhoneNumberModal and UserProfile
- Improve code reusability and maintainability
